requirements.txt
streamlit
streamlit_autorefresh
requests
pandas
plotly
pyyaml